refactor(wiki): dissolve lib/fileview/ into views/wiki/FileView.tsx + lib/wiki.ts
lib/fileview/ mixed a full page view (FileView, single-consumer of
WikiPage.tsx) with plain wiki-data lib code (SWR hooks, pageTitle, DocEntry/
ListResponse), which is what made the name confusing. Moves FileView and its
page-only subcomponents to views/wiki/FileView.tsx, sibling to WikiPage.tsx,
matching the lib/${feature}/ (hooks/utils/types/api) vs views/${feature}/
(full page views) split used elsewhere. Merges the remaining hooks/types/
utils into lib/wiki.ts, which already held the same kind of wiki-data code.
